ai_summarizer/requirements.txt
flask
google-generativeai
python-dotenv
PyPDF2
youtube-transcript-api
